docker_app/requirements.txt
streamlit==1.29.0
boto3==1.33.11
streamlit-cognito-auth==1.2.0
pillow
numpy
streamlit-drawable-canvas
opencv-python
